[Refactor] 디렉토리 구조 변경
- 각 페이지에서만 필요한 함수, 컴포넌트는 _components로 작성 후 삽입
- (route) 와 api 로 구분
- 과도한 컴포넌트 기능 분리
