+++ src/app/(route)/favicon.ico
