+++ src/app/(route)/layout.tsx
@@ -1,9 +1,8 @@
 import type { Metadata } from "next";
 import "./globals.css";
-import Script from "next/script";
 import Providers from "@/hooks/useReactQuery";
-import AuthProvider from "@/components/Auth/AuthProvider";
 import Navigator from "@/components/sidebar/Navigator";
+import AuthProvider from "@/components/auth/authprovider";
 
 export const metadata: Metadata = {
   title: "쓰파인더",
@@ -25,12 +24,7 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="ko">
-      <head>
-        <Script
-          strategy="beforeInteractive"
-          src={`//dapi.kakao.com/v2/maps/sdk.js?appkey=${process.env.NEXT_PUBLIC_KAKAO_MAP_KEY}&autoload=false&libraries=services,clusterer,drawing`}
-        />
-      </head>
+      <head />
       <body className="flex flex-row w-screen h-[100svh]">
         <Navigator />
         <AuthProvider>
