+++ src/app/(route)/setting/page.tsx
@@ -1,0 +1,7 @@
+export default function SettingPage() {
+  return (
+    <div>
+      <h1>설정페이지</h1>
+    </div>
+  );
+}
