--- src/app/location/layout.tsx
@@ -1,7 +1,0 @@
-export default function LocationLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return <main className="relative flex-grow">{children}</main>;
-}
